Separate parsing implicits in a different class.
I think this is better code than what I had before. Sadly, this is still
inefficient. We are creating the same objects again and again which we
shouldn't need to. But, all tests fail when I use `val`s. I added TODOs
to explain what's going on.
